requirements.txt
kivy
pytz
requests
Pillow
beautifulsoup4
ddddocr
python-dotenv
